+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to `laravel-sevdesk-api` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0 - 2021-12-05
 
 - initial release
+- implemented basic api handler
+- implemented sevdesk contact api